Use lexer term instead and merge inner+outer tokenizer
