+++ lib/index.js
@@ -1,11 +1,17 @@
 const {read} = require("./file")
 
-const tokenize = require("./tokenizer/index")
+const lexer = require("./tokenizer/lexer")
 const parser = require("./parser/index")
 
 module.exports = {
+  /**
+   * a wrapper for the lexer and parser
+   * 
+   * @param {string} path path to a markdown file
+   * @returns a html string
+   */
   renderer: (path) => {
     const fileContent = read(path).split("\n")
-    return parser(tokenize(fileContent))
+    return parser(lexer(fileContent))
   }
 }
