+++ lib/parser/index.js
@@ -47,7 +47,7 @@ const parser = (tokens) => {
       } else {
         cbHtml = cbHtml.replace("%class", "")
       }
-      parsedContent.push(cbHtml.replace("%s", token.tokens.value))
+      parsedContent.push(cbHtml.replace("%s", token.tokens[0].value))
       continue
     }
