+++ lib/regex.js
@@ -53,5 +53,6 @@ module.exports = {
     image: /^(?<indent>[\s]*)!\[(?<alt>.*)\]\((?<url>.*)\)/g
   },
     
-  PARAGRAPH_REGEX: /(\*\*)(?<bold>[^**]+)(\*\*)|`(?<code>[^`]+)`|(~~)(?<strike>[^~~]+)(~~)|\[(?<linkTitle>.*)]\((?<linkHref>.*)\)|(?<normal>[^`*[~]+)|\*(?<italic>[^**]+)\*/g
+  PARAGRAPH_REGEX: /(\*\*)(?<bold>[^**]+)(\*\*)|`(?<code>[^`]+)`|(~~)(?<strike>[^~~]+)(~~)|\[(?<linkTitle>.*)]\((?<linkHref>.*)\)|(?<normal>[^`*[~]+)|\*(?<italic>[^**]+)\*/g,
+  TABLE_REGEX: /^(?<indent>\s*)\|(?<value>[^\n\r]+)\|/g
 }
