+++ lib/tokenizer/lexer.js
@@ -1,0 +1,118 @@
+const { REGEX, findRegex, PARAGRAPH_REGEX } = require("../regex.js")
+const TOKENS = require("./tokens.js")
+
+const lexer = (lines) => {
+  // parse every line to the respective markdown tokens
+  const lexerContent = []
+  for(let i=0; i<lines.length; i++) {
+    const lineToParse = lines[i].replace("\r", "")
+    
+    // for paragraphs, we need to parse the line with the PARAGRAPH_REGEX
+    const beforeCompare = lexerContent.length
+    
+    // new line parsing
+    if(["", "\n"].includes(lineToParse)) {
+      if (beforeCompare === 0) continue // cannot be added at the top of the content
+      // if there are multiple new lines in a row,
+      // only single new line is added to the lexerContent
+      if (lexerContent[beforeCompare-1].type === TOKENS.newLine) continue
+      lexerContent.push({
+        type: TOKENS.newLine,
+      })
+      continue
+    }
+    
+    // check for the codeblock
+    if (lineToParse.startsWith("```")) {
+      // find the following "```"
+      let lineNumber = i + 1
+      while(lines.length > lineNumber && !lines[lineNumber].startsWith("```")) {
+        lineNumber++
+      }
+      const codeBlock = lines.slice(i+1, lineNumber)
+        .map(line => line.replace("\r", ""))
+        .join("\n")
+        // calculate the codeblock tokens here
+      const codeLanguage = lineToParse.substring(3)
+      lexerContent.push({
+        type: TOKENS.codeBlock,
+        tokens: [{
+          value: codeBlock
+        }],
+        language: codeLanguage || null
+      })
+      i = lineNumber
+      continue
+    }
+    
+    // look for the every regex in the REGEX object
+    for (let key in REGEX) {
+      const regex = REGEX[key]
+      let match = lineToParse.match(regex)
+
+      if (match) {
+        match = regex.exec(lineToParse)
+
+        const indent = match.groups.indent
+        delete match.groups.indent
+
+        const lexerData = {
+          type: TOKENS[key],
+          indent,
+          tokens: match.groups,
+          raw: lineToParse
+        }
+
+        if ([TOKENS.newLine, TOKENS.codeBlock, TOKENS.comment, TOKENS.image].includes(key)) {
+          // newline does not have inner tokens
+          // codeblock, comment, image doesn't need to be checked
+          lexerContent.push(lexerData)
+          break
+        }
+
+        if (key === TOKENS.heading) {
+          lexerData.level = match.groups.level.length
+          delete match.groups.level
+          lexerData.tokens = findRegex(PARAGRAPH_REGEX, match.groups.value)
+          lexerContent.push(lexerData)
+          break
+        }
+
+        if (key === TOKENS.quote) {
+          // if quote type, first we need to find the quote depth
+          // then we need to find the inner tokens
+          // then we need to add the quote depth to the inner tokens
+          // then we need to add the inner tokens to the parsed line
+          let depth = 0
+          let quoteLine = lexerData.tokens.value
+          while(quoteLine.trim().startsWith(">")) {
+            depth++
+            quoteLine = quoteLine.substring(1).trim()
+          }
+          // now the quoteline is pure raw text
+          // we can check it with the pqragraph regex
+          lexerData.depth = depth
+          lexerData.tokens = findRegex(PARAGRAPH_REGEX, quoteLine)
+          lexerContent.push(lexerData)
+          break
+        }
+
+        lexerData.tokens = findRegex(PARAGRAPH_REGEX, match.groups.value)
+        lexerContent.push(lexerData)
+        break
+      }
+    }
+    // if noting was found, we submit the content as paragraph for deep tokenization
+    if (beforeCompare === lexerContent.length) {
+      lexerContent.push({
+        type: TOKENS.paragraph,
+        tokens: findRegex(PARAGRAPH_REGEX, lineToParse),
+        raw: lineToParse // just \r is removed,
+      })
+    }
+    
+  }
+  return lexerContent
+}
+
+module.exports = lexer
